adicionado funcionalidades para deletar todas as tasks, inclusive impedindo a funcao de ser executada caso nao existam tasks. getTotalPercentProgress ativado ao init para atualizar o numero do progresso ao iniciar

+++ src/app/domain/task-list/components/tasks/task-title/task-title.component.ts
@@ -1,6 +1,9 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 
 import { TaskService } from '../../../services/task.service';
+import { ConfirmDialogComponent } from '../../widgets/confirm-dialog/confirm-dialog.component';
+import { ToastService } from './../../../services/toast.service';
 
 @Component({
   selector: 'app-task-title',
@@ -8,9 +11,26 @@ import { TaskService } from '../../../services/task.service';
   styleUrls: ['./task-title.component.scss'],
 })
 export class TaskTitleComponent {
-  constructor(private taskService: TaskService) {}
+  constructor(
+    private taskService: TaskService,
+    private toastService: ToastService,
+    public dialog: MatDialog
+  ) {}
 
   deleteAllTasks() {
-    this.taskService.deleteAllTasks();
+    if (this.taskService.tasks.length === 0) {
+      return this.toastService.showToastError('Não existem tasks para excluir');
+    }
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: 'Deseja realmente excluir todas as Tasks?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.taskService.deleteAllTasks();
+        this.toastService.showToastSucess('Tasks deletadas com sucesso');
+      }
+    });
   }
 }
